# docs/openapi.yaml
openapi: 3.1.0
info:
  title: RecallAI
  version: 0.1.0
paths:
...
  /api/v1/spaces/{space_id}/connections:
    get:
      tags:
      - spaces
      summary: Space Connections
      description: 'The caller''s own connections between memories in this Space.


        **Only their own.** A Space is where one person reads another''s rows, and this does

        not widen that: a connection is a judgement its author made about their own memories,

        so two members see the same Space with different graphs. Widening later is a decision

        somebody can take deliberately; un-showing people each other''s judgements is not.


        Two statements: the membership gate, then the edges. Both ends must be in the Space

        and both must still exist, so nothing here can render a card from outside it.'
      operationId: space_connections_api_v1_spaces__space_id__connections_get
      parameters:
      - name: space_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Space Id
      - name: recall_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Recall Session
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpaceConnectionsResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/connections/suggestions:
    get:
      tags:
      - connections
      summary: List Suggestions
      description: 'Edges nobody has accepted yet, strongest first.


        Per user rather than per memory: an edge found from a new capture''s side is a

        suggestion on *both* of its memories, so offering it per memory offers one decision

        twice. Declared above `/{connection_id}` -- a literal segment that could be read as a

        parameter has to be matched first.'
      operationId: list_suggestions_api_v1_connections_suggestions_get
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
            maximum: 100
            minimum: 1
          - type: 'null'
          title: Limit
      - name: recall_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Recall Session
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuggestionListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/connections/hubs:
    get:
      tags:
      - connections
      summary: List Hubs
      description: 'The caller''s most-connected memories, busiest first.


        What the page opens on when nobody named a memory. Reached from the nav, the only

        other option is a list to choose from -- and a list of the *newest* memories says

        nothing about which are worth opening, since the newest capture is usually the one

        with the fewest edges.


        An empty list is a real answer, not a 404: a vault with no connections yet has no hubs

        and the page falls back to offering recent memories.


        Declared above `/{connection_id}` -- a literal segment that could be read as a

        parameter has to be matched first.'
      operationId: list_hubs_api_v1_connections_hubs_get
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
            maximum: 100
            minimum: 1
          - type: 'null'
          title: Limit
      - name: recall_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Recall Session
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HubListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/connections/for-item/{item_id}:
    get:
      tags:
      - connections
      summary: Neighbourhood
      description: 'One memory and everything connected to it, read from both ends.


        Two statements: the ownership gate on the focus, then the neighbourhood and its total

        together. Inferring "not yours" from an empty edge list would save the first one and

        would stop being an ownership check.


        Declared above `/{connection_id}` for the literal-segment reason above.'
      operationId: neighbourhood_api_v1_connections_for_item__item_id__get
      parameters:
      - name: item_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Item Id
      - name: relation
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/Relation'
          - type: 'null'
          title: Relation
      - name: include_suggested
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Include Suggested
      - name: limit
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
            maximum: 100
            minimum: 1
          - type: 'null'
          title: Limit
      - name: recall_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Recall Session
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NeighbourhoodResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/connections:
    post:
      tags:
      - connections
      summary: Create Connection
      description: 'Connect two of the caller''s own memories.


        **200, not 201, and never 409.** The pair is unique regardless of order, so connecting

        B to A when A to B already exists re-labels the one edge rather than failing --

        re-adding is a normal outcome, not an error, and `created` says which happened. A 409

        would make the obvious action look like a broken server.


        Both ids are checked against the caller *per item*, not per request: a body carrying

        two ids is exactly where a stranger''s is easiest to slip in.'
      operationId: create_connection_api_v1_connections_post
      parameters:
      - name: recall_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Recall Session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateConnectionRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateConnectionResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/connections/{connection_id}:
    patch:
      tags:
      - connections
      summary: Update Connection
      description: Re-label an edge, or change its note. An empty note clears it.
      operationId: update_connection_api_v1_connections__connection_id__patch
      parameters:
      - name: connection_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Connection Id
      - name: recall_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Recall Session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateConnectionRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectionRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - connections
      summary: Delete Connection
      operationId: delete_connection_api_v1_connections__connection_id__delete
      parameters:
      - name: connection_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Connection Id
      - name: recall_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Recall Session
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/connections/{connection_id}/confirm:
    post:
      tags:
      - connections
      summary: Confirm Connection
      description: Accept a suggestion. Until this happens no lane that reads connections sees it.
      operationId: confirm_connection_api_v1_connections__connection_id__confirm_post
      parameters:
      - name: connection_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Connection Id
      - name: recall_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Recall Session
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectionRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/connections/{connection_id}/retype:
    post:
      tags:
      - connections
      summary: Retype Connection
      description: 'Ask a model how these two memories relate, and label the edge with its answer.


        The only connection route that spends a model call, which is why it is the only one

        behind a per-user hourly cap. 503 when typing is switched off, unconfigured, or the

        caller has used up that cap -- one answer for three causes, because they are the same

        thing from outside and separating them would report which of an operator''s settings is

        off.


        The answer is written to `ai_reason`, never to `note`: one is rendered as

        machine-written and the other as the owner''s own words.'
      operationId: retype_connection_api_v1_connections__connection_id__retype_post
      parameters:
      - name: connection_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Connection Id
      - name: recall_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Recall Session
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectionRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/connections/{connection_id}/dismiss:
    post:
      tags:
      - connections
      summary: Dismiss Connection
      description: 'Decline a suggestion, and never be offered that pair again.


        The row is kept rather than deleted: it is what stops the derivation re-proposing the

        pair. Connecting the two by hand later revives it.'
      operationId: dismiss_connection_api_v1_connections__connection_id__dismiss_post
      parameters:
      - name: connection_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Connection Id
      - name: recall_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Recall Session
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ConnectionOrigin:
      type: string
      enum:
      - user
      - ai
      title: ConnectionOrigin
      description: Who drew the edge. Never an authorization input -- only a thing to render.
    ConnectionRead:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        relation:
          $ref: '#/components/schemas/Relation'
        direction:
          type: string
          enum:
          - outgoing
          - incoming
          title: Direction
        origin:
          $ref: '#/components/schemas/ConnectionOrigin'
        status:
          $ref: '#/components/schemas/ConnectionStatus'
        score:
          anyOf:
          - type: number
          - type: 'null'
          title: Score
        note:
          anyOf:
          - type: string
          - type: 'null'
          title: Note
        ai_reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Ai Reason
        created_at:
          type: string
          format: date-time
          title: Created At
        memory:
          $ref: '#/components/schemas/VaultItemRead'
      type: object
      required:
      - id
      - relation
      - direction
      - origin
      - status
      - created_at
      - memory
      title: ConnectionRead
      description: One edge, as seen from the memory that was asked about.
...
    CreateConnectionRequest:
      properties:
        source_id:
          type: string
          format: uuid
          title: Source Id
        target_id:
          type: string
          format: uuid
          title: Target Id
        relation:
          $ref: '#/components/schemas/Relation'
          default: related_to
        note:
          anyOf:
          - type: string
            maxLength: 500
          - type: 'null'
          title: Note
      type: object
      required:
      - source_id
      - target_id
      title: CreateConnectionRequest
      description: 'Connect two of the caller''s own memories.


        Both ids are in the body rather than one in the path so the self-edge check is a

        schema validator: it answers 422 before any statement runs. The database carries the

        same rule as a CHECK constraint, which is defence in depth rather than the thing

        anyone sees.'
    CreateConnectionResponse:
      properties:
        connection:
          $ref: '#/components/schemas/ConnectionRead'
        created:
          type: boolean
          title: Created
      type: object
      required:
      - connection
      - created
      title: CreateConnectionResponse
      description: '`created` is false when the pair already had an edge, in either order.


        That is a 200 and not a 409: re-adding is a normal outcome, not an error. Somebody

        connects A to B from A''s page and, a week later, from B''s -- the obvious action must

        not look like a broken server.'
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    HubListResponse:
      properties:
        hubs:
          items:
            $ref: '#/components/schemas/HubRead'
          type: array
          title: Hubs
      type: object
      title: HubListResponse
      description: 'The busiest memories, busiest first.


        What `/connections` opens on when no memory is named. An empty list is a real answer

        -- a vault with no edges yet -- and the page falls back to offering recent memories

        rather than inventing a focus.'
    HubRead:
      properties:
        memory:
          $ref: '#/components/schemas/VaultItemRead'
        connection_count:
          type: integer
          title: Connection Count
      type: object
      required:
      - memory
      - connection_count
      title: HubRead
      description: One memory and how many things connect to it.
...
    NeighbourhoodResponse:
      properties:
        focus:
          $ref: '#/components/schemas/VaultItemRead'
        connections:
          items:
            $ref: '#/components/schemas/ConnectionRead'
          type: array
          title: Connections
        total:
          type: integer
          title: Total
          default: 0
      type: object
      required:
      - focus
      title: NeighbourhoodResponse
      description: Everything connected to one memory.
...
    Relation:
      type: string
      enum:
      - related_to
      - expands
      - supports
      - contradicts
      - inspired_by
      - depends_on
      - example_of
      - part_of
      title: Relation
      description: 'How one memory relates to another.


        Stored as `Text` rather than as a PG enum, like `SpaceRole` and

        `extraction_runs.status`: adding `refutes` one day must not need an `ALTER TYPE`

        inside a migration. The column is not the check -- values are re-validated on the way

        in, and an unrecognised one is read as `related_to`.


        `related_to` is the weakest claim in the list and doubles as this vocabulary''s

        "Other". A derived edge is always this one: a cosine distance says two memories are

        *close*, which is the only thing it can say -- two documents that flatly contradict

        each other are maximally close. Anything stronger is a person''s judgement, or a

        model''s, and is labelled as whichever it was.'
...
    SpaceConnectionRead:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        relation:
          $ref: '#/components/schemas/Relation'
        note:
          anyOf:
          - type: string
          - type: 'null'
          title: Note
        ai_reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Ai Reason
        created_at:
          type: string
          format: date-time
          title: Created At
        source:
          $ref: '#/components/schemas/VaultItemRead'
        target:
          $ref: '#/components/schemas/VaultItemRead'
      type: object
      required:
      - id
      - relation
      - created_at
      - source
      - target
      title: SpaceConnectionRead
      description: 'One edge between two memories in a Space, with both of its ends.


        Both, unlike `ConnectionRead`: a Space panel has no focus memory for the other end to

        be relative to, so neither end can be called "the other one".'
    SpaceConnectionsResponse:
      properties:
        connections:
          items:
            $ref: '#/components/schemas/SpaceConnectionRead'
          type: array
          title: Connections
        total:
          type: integer
          title: Total
          default: 0
      type: object
      title: SpaceConnectionsResponse
      description: '**Only the caller''s own edges**, both of whose memories are in this Space.


        A Space is the one place somebody reads another person''s rows, and this deliberately

        does not widen that. A connection is a judgement its author made -- "A contradicts B"

        is an opinion, not a fact about the Space -- so two members looking at the same Space

        see different graphs. That is honest, and it is the reversible direction: widening

        later is easy, un-showing people each other''s judgements is not.'
...
    SuggestionListResponse:
      properties:
        suggestions:
          items:
            $ref: '#/components/schemas/SuggestionRead'
          type: array
          title: Suggestions
        total:
          type: integer
          title: Total
          default: 0
      type: object
      title: SuggestionListResponse
    SuggestionRead:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        relation:
          $ref: '#/components/schemas/Relation'
        score:
          anyOf:
          - type: number
          - type: 'null'
          title: Score
        ai_reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Ai Reason
        created_at:
          type: string
          format: date-time
          title: Created At
        source:
          $ref: '#/components/schemas/VaultItemRead'
        target:
          $ref: '#/components/schemas/VaultItemRead'
      type: object
      required:
      - id
      - relation
      - created_at
      - source
      - target
      title: SuggestionRead
      description: 'An edge nobody has accepted yet, with both of its memories.


        Both ends, unlike `ConnectionRead`: a suggestion is read from an inbox rather than

        from one memory''s page, so there is no "the memory you asked about" to be the other

        end of.'
...
    UpdateConnectionRequest:
      properties:
        relation:
          anyOf:
          - $ref: '#/components/schemas/Relation'
          - type: 'null'
        note:
          anyOf:
          - type: string
            maxLength: 500
          - type: 'null'
          title: Note
      type: object
      title: UpdateConnectionRequest
      description: 'Re-label or re-caption an edge.


        `None` means "leave it alone"; the empty string is how a note is cleared. A nullable

        field with two meanings for null is a field nobody can use correctly -- the same rule

        `UpdateSpaceRequest` follows for `icon` and `accent`.'
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: recall_session
      description: Signed JWT session cookie set by /api/v1/auth/google/callback.